Added a safe to use save plot function
Limits saved animation to 200 frames to prevent python from taking too much system resources.
Can either show plot, or save plot. Attempting both causes unexpected behavior.
This is because the state of the propagation of orbit is conserved between the two functions.
One can only start where the other left off. Need to change the implementation of the propagation to fix this.